requirements.txt
streamlit==1.49.0
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain==0.3.27
openai==1.102.0
faiss-cpu==1.7.4
tiktoken==0.11.0
altair==5.5.0